+++ IsometricEngine/Page_Editor.h
@@ -4,15 +4,13 @@
  
 class Page_Editor : public page {
 
-	
 
 public:
 	void Create_item();
 	void Modify_item();
 	void ItemDB();
 	void SeeItemDataBase();
-	void  Update(float dt);
+	void  Update(double dt);
 	virtual void init(int Width, int Height, Game* g);
  
-
 };
